requirements.txt
langchain==0.0.167
python-dotenv==1.0.0
